Add the full template renderer with templates
Big commit! The rendering and the templates has been implemented. Now
basically all the basic ingredients in the code are complete. And the
camera module has been touch a little to return the location and focus
of camera in numpy array as well for further usage. Also the default
options file has been changed to add a few new options, like for
background.